Fix: eslint config and move component
