+++ src/layouts/Default/index.tsx
@@ -1,7 +1,7 @@
 import type { FC } from 'react'
 import { useContext } from 'react'
 
-import { Loader } from '~/components/Loader'
+import { Loader } from '~/components/shared/Loader'
 import { AuthContext } from '~/contexts/AuthContext'
 
 import type { LayoutProps } from '../types'
